+++ outputs.tf
@@ -1,6 +1,31 @@
-output "dns_name" {
+output "alb_name" {
+  description = "The name of the load balancer."
+  value       = var.name
+}
+
+output "alb_id" {
+  description = "The ID of the load balancer."
+  value       = module.alb.this_lb_id
+}
+
+output "alb_arn" {
+  description = "The ARN of the load balancer."
+  value       = module.alb.this_lb_arn
+}
+
+output "alb_dns_name" {
   description = "The DNS name of the load balancer."
-  value       = module.alb.dns_name
+  value       = module.alb.this_lb_dns_name
+}
+
+output "alb_arn_suffix" {
+  description = "ARN suffix of our load balancer - can be used with CloudWatch."
+  value       = module.alb.this_lb_arn_suffix
+}
+
+output "alb_zone_id" {
+  description = "The zone_id of the load balancer to assist with creating DNS records."
+  value       = module.alb.this_lb_zone_id
 }
 
 output "http_tcp_listener_arns" {
@@ -23,21 +48,6 @@ output "https_listener_ids" {
   value       = module.alb.https_listener_ids
 }
 
-output "load_balancer_arn_suffix" {
-  description = "ARN suffix of our load balancer - can be used with CloudWatch."
-  value       = module.alb.load_balancer_arn_suffix
-}
-
-output "load_balancer_id" {
-  description = "The ID and ARN of the load balancer we created."
-  value       = module.alb.load_balancer_id
-}
-
-output "load_balancer_zone_id" {
-  description = "The zone_id of the load balancer to assist with creating DNS records."
-  value       = module.alb.load_balancer_zone_id
-}
-
 output "target_group_arns" {
   description = "ARNs of the target groups. Useful for passing to your Auto Scaling group."
   value       = module.alb.target_group_arns
@@ -53,11 +63,14 @@ output "target_group_names" {
   value       = module.alb.target_group_names
 }
 
-output "load_balancer_security_group_id" {
-  description = "Security Group ID associated to the Load Balancer"
-  value       = aws_security_group.alb_sg.id
+output "target_group_attachments" {
+  value = aws_lb_target_group_attachment.alb_tg_attachment
 }
 
-output "load_balancer_target_groups_backend_port" {
-  value = var.target_groups.*.backend_port
+output "alb_route53_records" {
+  value = aws_route53_record.alb_record
+}
+
+output "alb_security_groups" {
+  value = module.security_group
 }