+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v2.0.0 - 2021-02-05
+
+### Changed
+
+*BREAKING CHANGES:*
+
+- Update to the latest version of ALB module (5.0+)
+- Use security-group module instead of resource
+- Remove log bucket creation
+- Add listener rules for protection
+- Refactor target groups
+- Add example in README
+
 ## v1.2.0 - 2020-11-26
 
 ### Changed